requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
gevent==22.10.2
greenlet==2.0.1
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
websocket-client==1.4.2
zope.event==4.5.0
zope.interface==5.5.1
